feat: Added new map, added author time and best time.
Also added new map that has the finish volume right by the start, used
for testing.

+++ game/src/level/level.odin
@@ -20,5 +20,5 @@ Level :: distinct struct {
 	start_velocity: spat.Vector,
 
 	// Stats
-	author_time: f32
+	author_time: f64
 }
